Transfer index names onto extracted data
Works for both run (via set_index) and state (via the index
argument). The data is stored in the R6 class as "index_names"
as that's slightly easier than keeping a copy of the R object
or doing a conversion between something like
std::vector<std::string> on every run. We can move it into the
C++ code later if we want though.