Allow setting state with index
This exists already in mode and is actually very useful, even though
we've removed the need for it in mode (this existed to support
setting the stochasic variables before we implemented direct
stochasic simulation). Getting this working on the GPU is a bit of
a faff and I'm feeling we might defer that until after the merge is
complete because there will likely be a few other GPU tweaks.

The use case for this soon is to support zeroing some variables
periodically, so that we can turn cumulative measurements into daily
measurements, as pomp does.  But also just to smooth over the difference
with mode of course.