Remove the call to std::min from binomial generator.
`std::min` is not available when building for CUDA, and using it leads
to undefined results. The `math.h` functions are available, but they
aren't overloaded by type so it would add some complexity to pick the
right one based on `int_type`. Using a ternary operator is the simplest
alternative.

This bug had been unnoticed because of the `__nv_exec_check_disable__`
annotation that is used throughout the file. I've now removed the
annotation to avoid future bugs.

I've also made the error formatting conditional on `!__CUDA_ARCH__`,
since `sprintf` isn't available either, which was causing warnings.