Support gamma on CUDA.
This worked out of the box simply by adding the right annotations. I've
tested this on all three parts of the domain (`shape < 1`, `shape == 1`
and `shape > 1`) and get results that are equivalent on the CPU and GPU
(given the same starting seed).

Oddly enough I do get very occasional insignificant floating point
differences between the two, despite using doubles on both. I suspect
this is down to differences in the some of the math functions. By
comparison, I get no difference at all on the uniform random number
generator.